refactor: bump version, move CLI entry point to cli.py and update console script path
